# tests/audit/golden/openapi_normalised.json
{
  "components": {
    "schemas": {
...
      "BenchReport": {
        "additionalProperties": false,
        "description": "Complete benchmark report payload consumed by the web app.",
        "properties": {
          "analyzed": {
            "items": {
              "$ref": "#/components/schemas/Featured"
            },
            "title": "Analyzed",
            "type": "array"
          },
          "baselineSolverId": {
            "default": "lmfit",
            "description": "Solver id whose median runtime defines the speedup baseline (x1.0). Was implicitly \"lmfit\" through hardcoded engine lookups (`by_name[\"lmfit\"]`) and a hardcoded web label; now an explicit contract field so adding a fourth backend or retiring lmfit doesn't ripple across every consumer. Additive 1.0 -> 1.1 bump: old payloads without this field default here.",
            "title": "Baselinesolverid",
            "type": "string"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryMeta"
            },
            "title": "Categories",
            "type": "array"
          },
          "gitBranch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Git branch name at benchmark run time. None when git is unavailable. Additive field.",
            "title": "Gitbranch"
          },
          "gitCommit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Git commit hash (short, $\\ge$7 chars) at benchmark run time. None when git is unavailable or the run is not in a git repo. Additive field \u2014 old payloads validate with None.",
            "title": "Gitcommit"
          },
          "inference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InferenceBlock"
              },
              {
                "type": "null"
              }
            ]
          },
          "manifest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManifestSignals"
              },
              {
                "type": "null"
              }
            ],
            "description": "Manifest headline numbers (mirrors `reports._headline`) \u2014 exposed so the web gate-state UI can render geomean / max $|\\Delta r^2|$ / win-rate / pinned ratio without a sidecar fetch. Additive 1.1 -> 1.2 bump; old payloads on disk validate against the bumped schema because Pydantic fills the default."
          },
          "panels": {
            "description": "Declarative panel registry consumed by the web PanelRenderer. Additive 1.2 -> 1.3 bump: old payloads validate with [] and the web app renders its explicit \"report predates panel specs\" empty state (no silent fallback to a hardcoded layout).",
            "items": {
              "$ref": "#/components/schemas/PanelSpec"
            },
            "title": "Panels",
            "type": "array"
          },
          "runTimestampUnix": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Unix epoch seconds when the benchmark run started (int). None when not captured. Additive field.",
            "title": "Runtimestampunix"
          },
          "schemaVersion": {
            "default": "1.7",
            "title": "Schemaversion",
            "type": "string"
          },
          "solvers": {
            "items": {
              "$ref": "#/components/schemas/SolverMeta"
            },
            "title": "Solvers",
            "type": "array"
          },
          "suite": {
            "items": {
              "$ref": "#/components/schemas/SuiteCase"
            },
            "title": "Suite",
            "type": "array"
          },
          "trustBlock": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrustBlock"
              },
              {
                "type": "null"
              }
            ],
            "description": "V&V evidence block emitted by oracles.audit.runner. Additive \u2014 Pydantic fills None for pre-Plan-E payloads, so old runs still parse."
          }
        },
        "required": [
          "solvers",
          "categories",
          "analyzed",
          "suite"
        ],
        "title": "BenchReport",
        "type": "object"
      },
...
      "CategoryMeta": {
        "additionalProperties": false,
        "description": "Benchmark category definition (id, label, case count, hue token).",
        "properties": {
          "hue": {
            "title": "Hue",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "n": {
            "title": "N",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "label",
          "n",
          "hue"
        ],
        "title": "CategoryMeta",
        "type": "object"
      },
...
      "Featured": {
        "additionalProperties": false,
        "description": "One analyzed case rendered in full detail across all panels.\n\nPer-backend metrics live in one `profiles` map (keyed by solver id); the\nremaining fields are case-level (shared across backends). A report carries a\n*list* of these (``BenchReport.analyzed``), one per deep-dived case, selectable\nin the UI; the `id` is unique within the list.",
        "properties": {
          "Ngrid": {
            "items": {
              "type": "integer"
            },
            "title": "Ngrid",
            "type": "array"
          },
          "baseline": {
            "title": "Baseline",
            "type": "number"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "corr": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Corr",
            "type": "array"
          },
          "crossN": {
            "title": "Crossn",
            "type": "number"
          },
          "exprEdges": {
            "description": "Tied/shared-parameter expression edges (camelCase keys on the wire: targetNode / targetParam / expression). Empty for unconstrained cases. Additive-minor field.",
            "items": {
              "$ref": "#/components/schemas/ExprEdge"
            },
            "title": "Expredges",
            "type": "array"
          },
          "fixedParams": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "description": "Per-node fixed parameter names, e.g. {\"p0\": [\"center\"]} means node p0's center was held at its guess value during the fit. Empty for unconstrained cases. Additive-minor field \u2014 old payloads validate with {}.",
            "title": "Fixedparams",
            "type": "object"
          },
          "globalFit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GlobalFit"
              },
              {
                "type": "null"
              }
            ]
          },
          "guess": {
            "items": {
              "type": "number"
            },
            "title": "Guess",
            "type": "array"
          },
          "guessParams": {
            "items": {
              "$ref": "#/components/schemas/PeakACS"
            },
            "title": "Guessparams",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "modelFormula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "LaTeX formula string for the featured case's model shape. Sourced from PeakModel.formula_latex in the oracle registry. None when no formula is registered. Additive field.",
            "title": "Modelformula"
          },
          "modelSourceFile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Relative path to the Rust kernel source for this case's model, e.g. \"crates/spectrafit-models/src/gaussian.rs\". None for landscape/optfn cases without a dedicated kernel. Additive field \u2014 old payloads validate with None.",
            "title": "Modelsourcefile"
          },
          "multidim": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MultiDim"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nestedAdequacy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NestedAdequacy"
              },
              {
                "type": "null"
              }
            ],
            "description": "Nested-model adequacy V&V result for this case. Populated when the benchmark engine runs the nested-order oracle (Task 4.3). None for cases without a known generative order or when the oracle was not executed. Additive-minor field \u2014 old payloads without nestedAdequacy validate with None."
          },
          "noise": {
            "title": "Noise",
            "type": "number"
          },
          "paramNames": {
            "items": {
              "type": "string"
            },
            "title": "Paramnames",
            "type": "array"
          },
          "peaks": {
            "items": {
              "$ref": "#/components/schemas/Peak"
            },
            "title": "Peaks",
            "type": "array"
          },
          "profiles": {
            "additionalProperties": {
              "$ref": "#/components/schemas/BackendProfile"
            },
            "title": "Profiles",
            "type": "object"
          },
          "ref": {
            "items": {
              "type": "number"
            },
            "title": "Ref",
            "type": "array"
          },
          "runsSched": {
            "items": {
              "type": "integer"
            },
            "title": "Runssched",
            "type": "array"
          },
          "schedule": {
            "items": {
              "type": "integer"
            },
            "title": "Schedule",
            "type": "array"
          },
          "truth": {
            "items": {
              "$ref": "#/components/schemas/PeakACS"
            },
            "title": "Truth",
            "type": "array"
          },
          "x": {
            "items": {
              "type": "number"
            },
            "title": "X",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "x",
          "ref",
          "guess",
          "truth",
          "noise",
          "baseline",
          "profiles",
          "peaks",
          "paramNames",
          "corr",
          "Ngrid",
          "schedule",
          "runsSched",
          "crossN"
        ],
        "title": "Featured",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InferenceBlock": {
        "additionalProperties": false,
        "description": "Top-level inference attached to BenchReport (additive 1.3 -> 1.4).\n\n``calibration`` and ``speed_inference`` are additive-minor fields (Task 5.3):\nold payloads that lack them validate with ``None``; the gate axis is inert\nwhen ``None`` (no-pass-by-absence rule).",
        "properties": {
          "calibration": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CalibrationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/CaseInference"
            },
            "title": "Cases",
            "type": "array"
          },
          "config": {
            "$ref": "#/components/schemas/InferenceConfig"
          },
          "equivalence": {
            "items": {
              "$ref": "#/components/schemas/EquivalenceResult"
            },
            "title": "Equivalence",
            "type": "array"
          },
          "speedInference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpeedInferenceResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "winnerStability": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Winnerstability",
            "type": "object"
          }
        },
        "required": [
          "config",
          "cases",
          "equivalence",
          "winnerStability"
        ],
        "title": "InferenceBlock",
        "type": "object"
      },
...
      "ManifestSignals": {
        "additionalProperties": false,
        "description": "Headline numbers from ``manifest.json`` surfaced via the contract.\n\nMirrors the keys :func:`oracles.reports._headline` writes; ``pinned``\nis populated from :func:`oracles.reports.read_perf_baseline` (``None``\nwhen no baseline is currently pinned). Surfaced to the web so the\ngate-state UI can render real values instead of telling users to run\n``spc-bench show-baseline``. Additive 1.1 -> 1.2 bump \u2014 the field on\n:class:`BenchReport` is optional so old payloads on disk validate\nwithout going through ``migrate.py``.",
        "properties": {
          "gateState": {
            "anyOf": [
              {
                "enum": [
                  "pass",
                  "warn",
                  "fail"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Aggregate gate verdict. Single source of truth; the web UI must render this rather than recomputing from regression flags.",
            "title": "Gatestate"
          },
          "geomeanSpeedupVsBaseline": {
            "title": "Geomeanspeedupvsbaseline",
            "type": "number"
          },
          "harmonicMeanSpeedupVsBaseline": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Harmonic mean of per-case speedup values vs the baseline solver. Complements the geometric mean per Eeckhout (2024): the harmonic mean is the correct aggregate for equal-time comparisons and is always $\\le$ the geometric mean for positively-skewed speedup distributions. Additive field \u2014 old payloads on disk validate with None.",
            "title": "Harmonicmeanspeedupvsbaseline"
          },
          "maxAbsDeltaR2": {
            "title": "Maxabsdeltar2",
            "type": "number"
          },
          "nonfiniteDr2CaseIds": {
            "description": "Case ids whose $|\\Delta r^2|$ vs the baseline was non-finite (NaN/Inf). The accuracy gate fails when this is non-empty; carried as a list so _sanitize (which coerces non-finite floats) cannot erase the signal.",
            "items": {
              "type": "string"
            },
            "title": "Nonfinitedr2Caseids",
            "type": "array"
          },
          "pinned": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PinnedBaseline"
              },
              {
                "type": "null"
              }
            ]
          },
          "regressions": {
            "title": "Regressions",
            "type": "integer"
          },
          "sanitizedValuePaths": {
            "description": "JSONPath-ish locators ('$.suite[3].m.jax.r2') of non-finite floats the presentation sanitizer coerced to 0.0 when this artifact was written. Non-empty means some rendered 0.0 values are suppressions, not measurements. Carried as a list so _sanitize cannot erase the signal; additive field \u2014 old payloads validate with [].",
            "items": {
              "type": "string"
            },
            "title": "Sanitizedvaluepaths",
            "type": "array"
          },
          "saturatedCategories": {
            "description": "Categories where every supported backend hits $r^2 \\ge$ 0.999 on every case. Differential below the noise floor; UI should mark these explicitly.",
            "items": {
              "type": "string"
            },
            "title": "Saturatedcategories",
            "type": "array"
          },
          "spectrafitWinRate": {
            "title": "Spectrafitwinrate",
            "type": "number"
          }
        },
        "required": [
          "geomeanSpeedupVsBaseline",
          "maxAbsDeltaR2",
          "spectrafitWinRate",
          "regressions"
        ],
        "title": "ManifestSignals",
        "type": "object"
      },
...
      "PanelSpec": {
        "additionalProperties": false,
        "description": "One report panel as data: which series, which chart, which slot.\n\nThe web app renders these generically (PanelRenderer); adding a panel is\none record in ``oracles.panels.DEFAULT_PANELS``, not a new React block.\n``source`` keys into the web-side series-builder registry \u2014 it is a contract\nstring, pinned both sides by the default-panels fixture test.\nAdditive 1.2 -> 1.3 bump together with ``BenchReport.panels``.\n\nNote:\n    ``chart_kind`` is one of :data:`ChartKind`, mirroring the components\n    exported by ``web/src/charts/index.tsx`` one-to-one; the web-side\n    dispatch (``panelRegistry.tsx``) carries the mirror table and a\n    vitest sync test pins the two sides together. ``\"band\"`` maps to the\n    Line component's band-series mode.",
        "properties": {
          "chartKind": {
            "enum": [
              "line",
              "scatter",
              "ecdf",
              "violin",
              "box",
              "heatmap",
              "field",
              "lollipop",
              "dumbbell",
              "sparkline",
              "band"
            ],
            "title": "Chartkind",
            "type": "string"
          },
          "desc": {
            "default": "",
            "title": "Desc",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "layout": {
            "$ref": "#/components/schemas/PanelLayout"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "chartKind",
          "source"
        ],
        "title": "PanelSpec",
        "type": "object"
      },
...
      "SolverMeta": {
        "additionalProperties": false,
        "description": "Solver legend entry (id, label, and theme color tokens).\n\nRe-exported from oracles.bench_contract for BenchReport.solvers.",
        "properties": {
          "color": {
            "description": "Primary theme color token (CSS `var(--c-...)` reference, with an optional CSS fallback) used for this backend's main chart marks.",
            "title": "Color",
            "type": "string"
          },
          "id": {
            "description": "Stable backend/solver key (e.g. `spectrafit`, `lmfit`, `scipy-ls-trf`) used to join this entry against per-backend results elsewhere in the report.",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "description": "Human-readable display name shown in legends and tables.",
            "title": "Label",
            "type": "string"
          },
          "soft": {
            "description": "Muted/soft variant of `color` (CSS `var(--c-...-soft)` reference) used for secondary chart elements \u2014 bands, fills, and de-emphasized series \u2014 that must read as the same backend without competing with the primary marks.",
            "title": "Soft",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "color",
          "soft"
        ],
        "title": "SolverMeta",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Serves the frozen BenchReport contract (the report the web UI consumes).",
    "title": "SpectraFit Benchmark API",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/report/{run_id}": {
      "get": {
        "description": "Return the validated :class:`BenchReport` for the given *run_id*.",
        "operationId": "get_report_api_report__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A specific run's report by id."
      }
    },
...
    "/api/v1/report/{run_id}": {
      "get": {
        "description": "Return the validated :class:`BenchReport` for the given *run_id*.",
        "operationId": "get_report_api_v1_report__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A specific run's report by id."
      }
    },
...
  }
}